requirements.txt
autopep8==1.5.7
MarkupSafe==2.0.1
numpy==1.24.0
openai==0.27.2
pandas
pandas-stubs==1.2.0.35
pycodestyle==2.8.0
python-dotenv==0.19.2
requests==2.27.0
# tweepy==4.12.0
discord.py==2.1.0
PyYAML == 6.0
aiohttp == 3.8.1
async_lru==1.0.3
oauthlib==3.2.0
textblob==0.15.3
dataset == 1.6.0
# platform
better_profanity == 0.7.0
nltk
sqlalchemy==1.4.19
psycopg2-binary == 2.9.1
# twitter-stream.py
Django == 4.1.5
django-cors-headers == 3.13.0
djangorestframework == 3.14.0
django-redis == 5.2.0
asgiref == 3.6.0
gunicorn == 20.1.0
opencv-python == 4.6.0.66
Pillow == 8.4.0
scikit-image == 0.19.0
scikit-learn == 0.24.2